+++ crates/contract/src/lib.rs
@@ -13,7 +13,7 @@ mod eth_call;
 pub use eth_call::{CallDecoder, EthCall};
 
 mod error;
-pub use error::{Error, Result};
+pub use error::{Error, Result, TransportErrorExt, TryParseTransportErrorResult};
 
 mod event;
 pub use event::{Event, EventPoller};
