+++ crates/transport/Cargo.toml
@@ -25,7 +25,6 @@ workspace = true
 [dependencies]
 alloy-json-rpc.workspace = true
 alloy-primitives.workspace = true
-alloy-sol-types.workspace = true
 
 base64.workspace = true
 futures-utils-wasm.workspace = true
